P2c: extract afxdp/tx/transmit.rs (transmit + recycle cluster, 6 fns + TxError ~450 LOC)
Extracts the XSK TX-ring submit + per-frame recycle cluster from
tx/mod.rs into sibling tx/transmit.rs:

- TxError (enum) — pub(in crate::afxdp).
- recycle_cancelled_prepared_offset — pub(in crate::afxdp).
- recycle_prepared_immediately — pub(in crate::afxdp) (callers in
  cos/cross_binding.rs and worker.rs:1974).
- remember_prepared_recycle — pub(in crate::afxdp).
- transmit_batch — pub(in crate::afxdp).
- transmit_prepared_batch — pub(super) (sole caller is tx/mod.rs:220,
  private re-export 'use transmit::transmit_prepared_batch;' keeps
  that resolving).
- transmit_prepared_queue — pub(in crate::afxdp).

cos_queue_dscp_rewrite stays in tx/mod.rs — it's a CoS-side helper
called only by cos/queue_service.rs (6 sites), not a submit-site
primitive. Will move with the drain carve.

Imports for tx/transmit.rs: VecDeque, Ordering, frame::{apply_dscp_rewrite_to_frame,
decode_frame_summary, frame_has_tcp_rst}, monotonic_nanos, FastMap,
PreparedTxRecycle, PreparedTxRequest, TxRequest, BindingWorker,
tx_frame_capacity, TX_BATCH_SIZE, XdpDesc, super::rings::{maybe_wake_tx,
reap_tx_completions}, super::stats::stamp_submits.

cargo build --bins clean (88 warnings, all pre-existing dead-code; 0
unused-imports new). cargo test --bins 865/0/2 (rolling baseline).

Plan: docs/pr/p2c-tx-transmit/plan.md (v3, addressing Codex r1+r2 findings).
